feat: trim
