--- .github/scripts/forward-port-missing/tests/conftest.py
@@ -1,5 +1,0 @@
-import sys
-from pathlib import Path
-
-__project_root__ = Path(__file__).resolve().parents[1]
-sys.path.append(str(__project_root__ / "src"))
