--- .github/scripts/forward-port-missing/tox.ini
@@ -1,25 +1,0 @@
-[tox]
-min_version = 4.0
-env_list =
-    py314t
-    py314
-    py313t
-    py313
-    py312
-    py311
-    py310
-    py39
-    format
-
-[testenv]
-uv_resolution = lowest-direct
-dependency_groups = dev
-commands =
-    pytest
-    mypy
-
-; We want to make sure we format fine with base config of ruff and black
-[testenv:format]
-dependency_groups = dev
-commands =
-    ruff format --check .
